refactor(tracks): extract Task 13 teardown driver into its own module
Task 13 added ~170 lines of teardown-specific gathering/driver logic
(gatherTeardownObservation, runBeginTeardown) directly to tracks.ts,
which two consecutive task reviews have now flagged as growing past
reviewability. Moves both functions verbatim (comments intact) into
cli/src/commands/watch/teardown-driver.ts, following the same
separation join.ts already established for the join/merge concern:
decision logic (decideTeardown) stays the sole authority in
protocol.ts, tracks.ts keeps orchestration (reconcileTracks,
executeRuntimeEffect's dispatch table) and production wiring
(defaultTrackRuntime), and the new module gets exactly the shared
driver helpers (applyProtocolToState/persist/refOf/withRef/
EffectRunResult) it needs, now exported from tracks.ts for reuse.
Pure mechanical move — no behavior change, no public surface change;
tsc/eslint/tests all still pass with the same 1527-test count.
